kv: move KeyValueDB from os/ to kv/, libos.a to libkv.a
Better code organization, and it will allow us to link less code into
the mon.
