+++ src/kv/KineticStore.cc
