+++ src/os/DBObjectMap.h
@@ -12,7 +12,7 @@
 #include <boost/scoped_ptr.hpp>
 
 #include "ObjectMap.h"
-#include "KeyValueDB.h"
+#include "kv/KeyValueDB.h"
 #include "osd/osd_types.h"
 #include "common/Mutex.h"
 #include "common/Cond.h"
