+++ src/os/Makefile.am
@@ -16,10 +16,8 @@ libos_a_SOURCES = \
 	os/HashIndex.cc \
 	os/IndexManager.cc \
 	os/JournalingObjectStore.cc \
-	os/LevelDBStore.cc \
 	os/LFNIndex.cc \
 	os/MemStore.cc \
-	os/KeyValueDB.cc \
 	os/KeyValueStore.cc \
 	os/ObjectStore.cc \
 	os/WBThrottle.cc
@@ -43,8 +41,7 @@ if WITH_LIBZFS
 libos_a_SOURCES += os/ZFSFileStoreBackend.cc
 endif
 
-libos_a_CXXFLAGS = ${AM_CXXFLAGS} -I rocksdb/include
-libos_a_LIBADD = libos_types.a
+libos_a_LIBADD = libos_types.a libkv.a
 
 if WITH_LTTNG
 libos_a_LIBADD += $(LIBOS_TP)
@@ -71,8 +68,6 @@ noinst_HEADERS += \
 	os/IndexManager.h \
 	os/Journal.h \
 	os/JournalingObjectStore.h \
-	os/KeyValueDB.h \
-	os/LevelDBStore.h \
 	os/LFNIndex.h \
 	os/MemStore.h \
 	os/KeyValueStore.h \
@@ -84,36 +79,11 @@ noinst_HEADERS += \
 	os/XfsFileStoreBackend.h \
 	os/ZFSFileStoreBackend.h
 
-if WITH_SLIBROCKSDB
-# build rocksdb with its own makefile
-# for some stupid reason this needs -fPIC...
-# PORTABLE=1 fixes the aarch64 build (-march=native doesn't work there)
-rocksdb/librocksdb.a:
-	cd rocksdb && EXTRA_CXXFLAGS=-fPIC PORTABLE=1 make -j$(shell nproc) static_lib
-libos_a_CXXFLAGS +=  -I rocksdb/include -fPIC
-libos_a_SOURCES += os/RocksDBStore.cc
-libos_a_LIBADD += rocksdb/librocksdb.a
-noinst_HEADERS += os/RocksDBStore.h
-endif
-
-if WITH_DLIBROCKSDB
-libos_a_SOURCES += os/RocksDBStore.cc
-libos_a_LIBADD += -lrocksdb
-noinst_HEADERS += os/RocksDBStore.h
-endif
-
 if WITH_LIBZFS
 libos_zfs_a_SOURCES = os/ZFS.cc
 libos_zfs_a_CXXFLAGS = ${AM_CXXFLAGS} ${LIBZFS_CFLAGS}
 noinst_LIBRARIES += libos_zfs.a
 noinst_HEADERS += os/ZFS.h
-endif
-
-if WITH_KINETIC
-libos_a_SOURCES += os/KineticStore.cc
-libos_a_CXXFLAGS += -std=gnu++11
-libos_a_LIBADD += -lkinetic_client -lprotobuf -lglog -lgflags libcrypto.a
-noinst_HEADERS += os/KineticStore.h
 endif
 
 endif # ENABLE_SERVER
