+++ src/os/newstore/NewStore.h
@@ -27,7 +27,7 @@
 #include "common/WorkQueue.h"
 #include "os/ObjectStore.h"
 #include "os/fs/FS.h"
-#include "os/KeyValueDB.h"
+#include "kv/KeyValueDB.h"
 
 #include "newstore_types.h"
