+++ src/test/ObjectMap/test_keyvaluedb_atomicity.cc
@@ -1,7 +1,7 @@
 // -*- mode:C++; tab-width:8; c-basic-offset:2; indent-tabs-mode:t -*-
 #include <pthread.h>
 #include "include/buffer.h"
-#include "os/KeyValueDB.h"
+#include "kv/KeyValueDB.h"
 #include <sys/types.h>
 #include <dirent.h>
 #include <string>
